requirements.txt
bcrypt==5.0.0
blinker==1.9.0
click==8.3.2
et_xmlfile==2.0.0
Flask==3.1.3
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
numpy==2.4.4
openpyxl==3.1.5
pandas==3.0.2
psycopg2-binary==2.9.11
python-dateutil==2.9.0.post0
six==1.17.0
Werkzeug==3.1.8
gunicorn==22.0.0
